RRG_TDV-main/RRG_TDV-main/stock_volume_alert/requirements.txt
streamlit>=1.36
pandas>=2.2
plotly>=5.22
requests>=2.32
websocket-client>=1.8

